move odom with range sensor to odom_generator.py to terrain robust altitude estimation
