+++ jsk_uav_forest_common/launch/challenge.launch
@@ -7,7 +7,13 @@
   </include>
 
   <!-- Odometry -->
-  <node pkg="jsk_uav_forest_common" type="odom_generator.py" name="odom_gen" output="screen" />
+  <node pkg="jsk_uav_forest_common" type="odom_generator.py" name="odom_gen" output="screen" >
+    <arg name="use_lidar" value="True" />
+    <arg name="use_guidance_vel" value="True" />
+    <arg name="lidar_tc" value="0.5" />
+    <arg name="lidar_noise_cut_thresh" value="0.5" />
+    <arg name="guidance_vel_weight" value="0.9" />
+  </node>
 
   <!-- perception + motion -->
   <include file="$(find jsk_uav_forest_motion)/launch/motion.launch">
