+++ jsk_uav_forest_motion/launch/motion.launch
@@ -15,8 +15,6 @@
   <node pkg="jsk_uav_forest_motion" type="circle_motion.py" name="circle_motion" output="screen" args="$(arg use_dji)">
     <param name="task_kind" value="$(arg task_kind)" />
     <param name="uav_odom_sub_topic_name" value="$(arg uav_odom_topic_name)" if="$(arg use_dji)"/>
-    <param name="use_lidar" value="True" if="$(arg use_dji)"/>
-    <param name="use_guidance_vel" value="True" if="$(arg use_dji)"/>
     <rosparam file="$(find jsk_uav_forest_motion)/config/circle_motion/circle_motion.yaml" command="load"/>
     </node>
