+++ jsk_uav_forest_motion/script/circle_motion.py
@@ -54,20 +54,11 @@ class CircleMotion:
         self.odom_update_flag_ = False
         self.uav_xy_pos_ = np.zeros(2)
         self.uav_z_pos_ = 0.0
-        self.uav_roll_ = 0.0
-        self.uav_pitch_ = 0.0
         self.uav_yaw_ = 0.0
         self.uav_yaw_old_ = 0.0
         self.uav_yaw_overflow_ = 0
         self.uav_accumulated_yaw_ = 0.0
 
-        self.lidar_update_flag_ = False
-        self.uav_lidar_z_ = 0.0
-        self.uav_lidar_z_old_ = 0.0
-
-        self.guidance_vel_update_flag_ = False
-        self.guidance_vel_z_diff_ = 0.0
-        self.guidance_vel_stamp_old_ = rospy.Time.now()
 
         self.cycle_count_ = 0
         self.tree_xy_pos_ = np.zeros(2)
@@ -79,8 +70,6 @@ class CircleMotion:
         self.target_pos_pub_topic_name_ = rospy.get_param("~target_pos_pub_topic_name", "uav_target_pos")
         self.uav_odom_sub_topic_name_ = rospy.get_param("~uav_odom_sub_topic_name", "ground_truth/state")
         self.tree_location_sub_topic_name_ = rospy.get_param("~tree_location_sub_topic_name", "tree_location")
-        self.lidar_sub_topic_name_ = rospy.get_param("~lidar_sub_topic_name", "/guidance/ultrasonic")
-        self.guidance_vel_sub_topic_name_ = rospy.get_param("~guidance_vel_topic_name", "/guidance/velocity")
         self.tree_detection_start_service_name_ = rospy.get_param("~tree_detection_start_service_name", "sub_control")
         self.task_start_service_name_ = rospy.get_param("~task_start_service_name", "task_start")
 
@@ -99,11 +88,6 @@ class CircleMotion:
         self.circle_y_vel_ = rospy.get_param("~circle_y_vel", 0.5)
         self.tree_detection_wait_ = rospy.get_param("~tree_detection_wait", 1.0)
         self.task_kind_ = rospy.get_param("~task_kind", 1) #1 yosen 2 honsen 3 kesshou
-        self.use_lidar_ = rospy.get_param("~use_lidar", False)
-        self.lidar_tc_ = rospy.get_param("~lidar_tc", 0.5) #0.0~1.0
-        self.lidar_noise_cut_thresh_ = rospy.get_param("~lidar_noise_cut_thresh", 0.5)
-        self.use_guidance_vel_ = rospy.get_param("~use_guidance_vel", True)
-        self.guidance_vel_weight_ = rospy.get_param("~guidance_vel_weight", 0.9)
 
         self.control_timer_ = rospy.Timer(rospy.Duration(1.0 / self.control_rate_), self.controlCallback)
 
@@ -112,24 +96,14 @@ class CircleMotion:
         self.target_pos_pub_ = rospy.Publisher(self.target_pos_pub_topic_name_, Odometry, queue_size = 10)
         self.odom_sub_ = rospy.Subscriber(self.uav_odom_sub_topic_name_, Odometry, self.odomCallback)
         self.tree_location_sub_ = rospy.Subscriber(self.tree_location_sub_topic_name_, PointStamped, self.treeLocationCallback)
-        self.lidar_sub_ = rospy.Subscriber(self.lidar_sub_topic_name_, LaserScan, self.lidarCallback)
-        self.guidance_vel_sub_ = rospy.Subscriber(self.guidance_vel_sub_topic_name_, Vector3Stamped, self.guidanceVelCallback)
         self.task_start_service_ = rospy.Service(self.task_start_service_name_, SetBool, self.taskStartCallback)
-
-        ###debug###
-        self.debug_pub_ = rospy.Publisher("/lidar_debug", Float32, queue_size = 10)
 
     def odomCallback(self, msg):
         self.odom_ = msg
         self.uav_xy_pos_ = np.array([msg.pose.pose.position.x, msg.pose.pose.position.y])
-        if self.use_lidar_ == True:
-            self.uav_z_pos_ = self.uav_lidar_z_
-        else:
-            self.uav_z_pos_ = msg.pose.pose.position.z
+        self.uav_z_pos_ = msg.pose.pose.position.z
         quaternion = np.array([msg.pose.pose.orientation.x, msg.pose.pose.orientation.y, msg.pose.pose.orientation.z, msg.pose.pose.orientation.w])
 
-        self.uav_roll_ = tf.transformations.euler_from_quaternion(quaternion)[0]
-        self.uav_pitch_ = tf.transformations.euler_from_quaternion(quaternion)[1]
         self.uav_yaw_ = tf.transformations.euler_from_quaternion(quaternion)[2]
 
         if self.odom_update_flag_ == True:
@@ -162,45 +136,7 @@ class CircleMotion:
             res.message = "Task does not start"
             rospy.loginfo("Task does not start")
         return res
-
-    def guidanceVelCallback(self, msg):
-        if self.guidance_vel_update_flag_ == False:
-            self.guidance_vel_update_flag_ = True
-            self.guidance_vel_stamp_old_ = msg.header.stamp
-            return
-        self.guidance_vel_z_diff_ = (-msg.vector.z * math.cos(self.uav_roll_) * math.cos(self.uav_pitch_)) * (msg.header.stamp - self.guidance_vel_stamp_old_).to_sec()
-        self.guidance_vel_stamp_old_ = msg.header.stamp
-
-    def lidarCallback(self, msg):
-        if self.use_guidance_vel_ == True:
-            if (self.odom_update_flag_ == False) or (self.guidance_vel_update_flag_ == False):
-                self.uav_lidar_z_old_ = self.uav_lidar_z_
-                self.lidar_update_flag_ = True
-                return
-        else:
-            if self.odom_update_flag_ == False:
-                self.uav_lidar_z_old_ = self.uav_lidar_z_
-                self.lidar_update_flag_ = True
-                return
-
-        lidar_val = msg.ranges[0] * math.cos(self.uav_roll_) * math.cos(self.uav_pitch_)
-
-        if self.use_guidance_vel_ == True:
-            self.uav_lidar_z_ = self.uav_lidar_z_old_ + self.guidance_vel_z_diff_ #update by guidance velocity integration
-            if lidar_val > self.lidar_noise_cut_thresh_: #if reliable
-                self.uav_lidar_z_ = lidar_val * (1.0 - self.guidance_vel_weight_) + self.uav_lidar_z_ * self.guidance_vel_weight_ #fusion
-            self.uav_lidar_z_ = self.lidar_tc_ * self.uav_lidar_z_old_ + (1.0 - self.lidar_tc_) * self.uav_lidar_z_ #LPF
-        else:
-            if lidar_val > self.lidar_noise_cut_thresh_: #if reliable
-                self.uav_lidar_z_ = self.lidar_tc_ * self.uav_lidar_z_old_ + (1.0 - self.lidar_tc_) * lidar_val #LPF
-            else:
-                pass #no update
-        self.uav_lidar_z_old_ = self.uav_lidar_z_ #update
-
-        pub_msg = Float32()
-        pub_msg.data = self.uav_lidar_z_
-        self.debug_pub_.publish(pub_msg)
-
+        
     def isConvergent(self, frame, target_xy_pos, target_z_pos, target_yaw):
         if frame == self.GLOBAL_FRAME_:
             delta_pos = np.array([target_xy_pos[0] - self.uav_xy_pos_[0], target_xy_pos[1] - self.uav_xy_pos_[1], target_z_pos - self.uav_z_pos_])
@@ -286,8 +222,6 @@ class CircleMotion:
 
     def controlCallback(self, event):
         if (not self.odom_update_flag_) or (not self.task_start_):
-            return
-        if self.use_lidar_ and (not self.lidar_update_flag_):
             return
 
     #navigation
